requirements.txt
bleach==2.0.0
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
decorator==4.1.2
Django==1.11.5
django-debug-toolbar==1.8
django-extensions==1.9.1
entrypoints==0.2.3
html5lib==0.999999999
idna==2.6
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.0.0
olefile==0.44
pandocfilters==1.4.2
pickleshare==0.7.4
pilkit==2.0
Pillow==4.2.1
prompt-toolkit==1.0.15
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
pyzmq==16.0.2
requests==2.18.4
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.2
